Port prices command to new command system
